BugID:19120134:Add http parser implementations

+++ network/http/include/http_wrapper.h
@@ -17,7 +17,7 @@ int httpc_wrapper_ssl_connect(int socket, const struct sockaddr *name, socklen_t
 int httpc_wrapper_connect(int socket, const struct sockaddr *name, socklen_t namelen);
 int httpc_wrapper_send(int socket, const void *data, uint16_t size, int flags);
 
-typedef int (* httpc_wrapper_recv_fn_t)(httpc_t *http_session, void *data, uint16_t size);
+typedef int (* httpc_wrapper_recv_fn_t)(httpc_t *http_session, void *data, int32_t size);
 int httpc_wrapper_register_recv(httpc_t *httpc, httpc_wrapper_recv_fn_t);
 
 #endif  /* HTTP_CLIENT_WRAPPER_H */
